server_recommendations_26/fetch_wardrobe/requirements.txt
fastapi
uvicorn
python-dotenv
azure-cosmos
